product card wip

+++ lib/api/items.ts
@@ -5,7 +5,6 @@ import {
   defaultPaginatedResponse,
   PaginatedResponse,
 } from "./types/itemTypes/PaginatedResponse";
-import qs from "qs";
 import { buildQueryString } from "../utils";
 
 export const itemService = {
@@ -25,8 +24,37 @@ export const itemService = {
     );
   },
 
+  getAllForAdmin: async (
+    headers?: Record<string, string>,
+    options?: PaginationOptions
+  ): Promise<ApiResponse<PaginatedResponse<Item>>> => {
+    const queryString = buildQueryString(options);
+
+    return request(
+      api.get(`/admin/items?${queryString}`, {
+        headers: {
+          ...headers,
+        },
+      }),
+      defaultPaginatedResponse
+    );
+  },
+
   getOne: async (id: string): Promise<ApiResponse<Item>> =>
     request(api.get(`/items/${id}`), {} as Item),
+
+  getOneForAdmin: async (
+    id: string,
+    headers?: Record<string, string>
+  ): Promise<ApiResponse<Item>> =>
+    request(
+      api.get(`/admin/items/${id}`, {
+        headers: {
+          ...headers,
+        },
+      }),
+      {} as Item
+    ),
 
   getBySlug: async (
     slug: string,
@@ -46,7 +74,7 @@ export const itemService = {
     headers?: Record<string, string>
   ): Promise<ApiResponse<Item>> =>
     request(
-      api.post("/items", data, {
+      api.post("/admin/items", data, {
         headers: {
           "Content-Type":
             data instanceof FormData
@@ -64,7 +92,7 @@ export const itemService = {
     headers?: Record<string, string>
   ): Promise<ApiResponse<Item>> =>
     request(
-      api.patch(`/items/${id}`, data, {
+      api.patch(`/admin/items/${id}`, data, {
         headers: {
           "Content-Type":
             data instanceof FormData
@@ -80,5 +108,5 @@ export const itemService = {
     id: number,
     headers?: Record<string, string>
   ): Promise<ApiResponse<null>> =>
-    request(api.delete(`/items/${id}`, { headers }), null),
+    request(api.delete(`/admin/items/${id}`, { headers }), null),
 };
